# Make getJacobian simpler and faster