@marioprats 2023-09-27
Thanks for the reviews.
While going again through the code adding comments I realized of an issue when the joint group origin doesn't match the URDF origin. Added a test for it and should be ok now.